==> URLPhis_Model/1.URL.ipynb <==
# 1.URL creation

## Selecting from 5000 of Phishing URLs and 5000 of legitimate URLs


*   Sources: 

     http://data.phishtank.com/data/online-valid.csv 
     https://www.unb.ca/cic/datasets/url-2016.html 



#importing required package 
import pandas as pd

### 1. Selecting and Creating Phishing URLs

data0 = pd.read_csv("./1.Phishing.csv")
data0.head() 

data0.shape

phishurl = data0.sample(n = 5000, random_state = 12).copy()
phishurl = phishurl["url"]
phishurl = phishurl.reset_index(drop=True)
phishurl.head()

# Storing the extracted legitimate URLs fatures to csv file
phishurl.to_csv('./3.phishurl.csv', index= False)

### 1. Selecting and Creating legitimate URLs

data1 = pd.read_csv("./2.Legitimate.csv")
data1.columns = ['URLs']
data1.head()

#Collecting 5,000 Legitimate URLs randomly
legiurl = data1.sample(n = 5000, random_state = 12).copy()
legiurl = legiurl.reset_index(drop=True)
legiurl.shape

# Storing the extracted legitimate URLs fatures to csv file
legiurl.to_csv('./4.legitimateurl.csv', index= False)
